Update patch set 2
Patch Set 2:

Hi Mark,

First of all, these 2 scripts are releng specific and should not be located in project repos. Please abandon your patch and we do what's necessary in releng repo in common way like the other scenarios people are working on doing.

When it comes to testing ovsnfv with fuel; this should be one of the scenarios on the Wiki. I see an ovs scenario on below page. Please check the page, fix what is missing and take the ownership of the scenario.

https://wiki.opnfv.org/brahmaputra_testing_page#test_scenarios_for_brahmaputra

You need to talk to Fuel team for the details regarding scenario definition and configuration file. 
Once this part is clarified, needed changes will be reflected to Jenkins jobs and releng scripts.
Once this is done, scenario dealing with ovsnfv will be executed by a daily job in scope of fuel so you don't need to fix this separately.

Let me know if this clarifies some of the questions you might have.

Patch-set: 2