Update patch set 2
Patch Set 2:

(2 comments)

How are we going to reconcile between two builds:  one for fuel and one for Apex. I am still a little confused by the fuel build process. Is this a script that is called by fuel or is it intended to be called directly by the ovsnfv build.sh? I am asking because I am concerned tha the fuel build may require a different server then the RPM build.

I am not giving this a -1 though.

Patch-set: 2