Update patch set 2
Patch Set 2:

Thomas,

Fuel and Apex builds are done by scripts fuel/ci/build.shThomas,

Fuel and Apex builds are done by scripts fuel/ci/build.sh apex/ci/build.sh

https://gerrit.opnfv.org/gerrit/gitweb?p=fuel.git;a=tree;f=ci
https://gerrit.opnfv.org/gerrit/gitweb?p=apex.git;a=tree;f=ci

In Fuel case, your stuff gets included in the ISO as plugin which is fixed by Jonas I think.
You need to talk to apex guys to do the similar thing for apex; for example ovsnfv builds an rpm, stores it on artifacts.opnfv.prg and that rpm gets pulled by apex build script during build time or in a similar way.

Both of these build.sh scripts are executed by different jenkins jobs specific to apex and fuel.

We already have build servers for fuel and apex so this shouldn't be a concern for you.

When it comes to deployment scenarios; Jonas already explained how it is intended to be done for fuel.
ovsnfv plugin is in the ISO and the fuel deployment will be run by ovsnfv scenarios, activating your plugin.

In short; you either build your stuff and store it in artifacts for others to consume or others build your stuff in scope of their build process.

Patch-set: 2