Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

I share the concern that this change now requires a Debian based build machine but, since you are building a Debian package, that's probably not fatal.

A bigger concern is that you now can't recreate a prior binary.  If the repo that `aptitude' is using is updated with newer versions of a package then we won't be able to get the older version that was used when creating release binaries.  Not being able to re-create a release seems like a bad idea.

I've also added some comments inline about the script itself.

Patch-set: 1
Reviewer: Gerrit User 1453 <1453@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1