Update patch set 1
Patch Set 1:

(3 comments)

Hi Don,

thanks for script comments - all very sensible, I'll port them to the main branch.

> A bigger concern is that you now can't recreate a prior
> binary.

This is also the case previous to this patch. 

This patch is to solve the problem of .debs in the archive.ubuntu.com being updated **and the previous .deb being removed** thereby breaking the plugin build and also the .iso build. (This happened a few weeks ago with libxml2_2.9.1).

I can think of one other solution to avoid this: pull down the required .debs from archive.ubuntu.com and store them somewhere - in the plugin repo perhaps.

Patch-set: 1