Update patch set 4
Patch Set 4: Code-Review+2

I am going to be impolite and merge this change. Several of you reviewed the previous change that added sudo. This patch is pretty innocuous but also necessary for successful clearing of the domain so next invocation will work. Next patch series will be OVS with DPDK.

Patch-set: 4
Label: Code-Review=+2