Update patch set 2
Patch Set 2:

Hi Tom,

afaik any comment with reverify in it should trigger the ci/build.sh script again. I don't know if there is a difference between a daily build, merge build etc - afaik the releng scrips are all set up to invoke just ci/build.sh ? 

It would be good if it could be parameterised so that we could do different tests depending on what part of the tree is modified - currently there is nothing there to build the fuel plugin for instance. And fuel plugin changes invoke the rpm build which isn't really required. 

Could be something for the rel-B retrospective.

Patch-set: 2