Update patch set 2
Patch Set 2:

Yes, but there is no way I know of to trigger a daily build. As for parameterized builds, I agree that they would be a good idea. I think though that RELENG project controls the parameters. They are not in our own repo. These changes have to be made with patches to RELENG.

Patch-set: 2