fuel plugin: Fix mechanism driver config
This fixes the yardstick ping test. And hopefully other y/s tests.

Add ovs to the mechanism driver search list so that neutron agent
continues to work with the kernel ovs on the non-compute nodes.

This enables the router on the control node and makes VMs accessible via
public ip addresses.

Change-Id: Ife08e63e8bf9358cc24eb610b3d7611484b3151f
Signed-off-by: Billy O'Mahony <billy.o.mahony@intel.com>
Reviewed-by: Jonas Bjurel <jonas.bjurel@ericsson.com>
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>
Reviewed-by: Tom Herbert <therbert@redhat.com>
Reviewed-by: Michal Skalski <mskalski@mirantis.com>
Reviewed-by: Stefan Berg <stefan.k.berg@ericsson.com>