Update patch set 1
Patch Set 1:

Will this mean that the specs will eventually be pulled into the OPNFV release documentation? Is that what we want especially if this stuff is work-in-progress? Completed specs could be sensible. Not a big issue we - could just also just mark the spec as DRAFT or WIP then it'd probably be ok in the release docs.

Patch-set: 1
Label: Verified=0