Update patch set 1
Patch Set 1:

I would think so.  
I'd focus on ensuring each of the URL's I will produce are presented well as they can be.  We can then link to each of them when needed and have the complete OVSNFV collection available by reference on the "Colorado Library"when ready

Patch-set: 1
Label: Code-Review=0